requirements.txt
aiogram~=3.10.0
python-dotenv~=1.0.1
asyncpg~=0.29.0